+++ handlers/routes_test.go
@@ -4,8 +4,6 @@ import (
 	"log"
 	"testing"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/gorilla/mux"
 	"github.com/stretchr/testify/assert"
 )
@@ -45,28 +43,6 @@ func TestAPIRoutes(t *testing.T) {
 			router := handler.SetUpRoutes()
 			got := GetRouteMap(router)[tt.endPoint]
 			assert.Equal(tt.endPoint, got, "got = %v, want = %v", got, tt.endPoint)
-		})
-	}
-}
-
-func TestAWSHandler_SdkWrapperAPI(t *testing.T) {
-	assert := assert.New(t)
-	t.Parallel()
-	tests := []struct {
-		name    string
-		apiName string
-	}{
-		{"Check IAMGetUserCountHandler", "GetUserCount"},
-		{"Check IAMGetUserIdentityHandler", "GetUserIdentity"},
-		{"Check IAMGetAliasesHandler", "GetAliases"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			handler := NewDefaultAWSHandler(false)
-			emptyCfg := aws.Config{}
-			noOpClient := iam.NewFromConfig(emptyCfg)
-			_, err := handler.SdkWrapperAPI(noOpClient, tt.apiName)
-			assert.Error(err, "No Err Occured with Empty Profile. err = %v ", err)
 		})
 	}
 }
